Refactoring and combining utility functions Reducing files Updating phpcs
