+++ includes/enqueue-scripts.php
@@ -1,4 +1,5 @@
 <?php
+if ( ! defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly
 
 /**
  * Loads scripts and styles for the webform admin page.
